requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
reportlab==4.0.4
xlsxwriter==3.1.9
Werkzeug==2.3.7
gunicorn==21.2.0
